+++ api_yamdb/api_yamdb/constants.py
@@ -10,6 +10,7 @@ MAX_SCORE = 10
 MAX_TEXT_LENGTH = 10000
 
 # users
+CONF_CODE_MAX_LENGTH = 128
 STR_MAX_LENGTH = 150
 EMAIL_MAX_LENGTH = 254
 FORBIDDEN_USERNAME = 'me'
